vuelta al size_t

+++ PS/utils/ft_nlstbigger.c
@@ -6,13 +6,13 @@
 /*   By: xmatute- <xmatute-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/29 15:26:32 by xmatute-          #+#    #+#             */
-/*   Updated: 2022/11/29 16:39:50 by xmatute-         ###   ########.fr       */
+/*   Updated: 2022/11/29 17:03:14 by xmatute-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
-int	ft_nlstbigger(t_nlist	*l, int n)
+size_t	ft_nlstbigger(t_nlist	*l, int n)
 {
 	size_t	i;
 
@@ -26,7 +26,7 @@ int	ft_nlstbigger(t_nlist	*l, int n)
 	return (i);
 }
 
-int	ft_nlstsmaller(t_nlist	*l, int n)
+size_t	ft_nlstsmaller(t_nlist	*l, int n)
 {
 	size_t	i;
